change validation result structure to include non-consensus errors

+++ packages/rs-dpp/src/version/mod.rs
@@ -3,10 +3,10 @@ use std::collections::HashMap;
 
 mod protocol_version_validator;
 
-pub const LATEST_VERSION: u32 = 1;
+pub const LATEST_VERSION: u64 = 1;
 
 lazy_static! {
-    pub static ref COMPATIBILITY_MAP: HashMap<u32, u32> = {
+    pub static ref COMPATIBILITY_MAP: HashMap<u64, u64> = {
         let mut m = HashMap::new();
         m.insert(1, 1);
         m
